+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - PAM integrations fixed for unlock on login
 - Nextcloud Client service startup fixes
+- Swayidle wait for Hyprland to start
 ## [0.2.0] - 2025-03-19
 
 ### Added